% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessible-color-scales.R
\docType{data}
\name{accessible-palettes}
\alias{accessible-palettes}
\alias{palette_okabeito}
\alias{palette_datawrapper}
\title{Accessible color palettes}
\format{
\code{palette_okabeito} is a character vector of 7 hex code colors.

\code{palette_datawrapper} is a character vector of 7 hex code colors.
}
\usage{
palette_okabeito

palette_datawrapper
}
\description{
Accessible color palettes are designed to be friendly to colorblind and
colorweak readers. Additionally, these palettes should be friendly to
grayscale printing.
}
\examples{
scales::show_col(palette_okabeito)
scales::show_col(palette_datawrapper)
}
\references{
Okabe M., & Ito, K. (2008). \emph{Color universal design (CUD): How to
make figures and presentations that are friendly to Colorblind people}.
Jfly. \url{https://jfly.uni-koeln.de/color/}

Muth, L. C. (2020, June 23). \emph{What to consider when visualizing
data for colorblind readers}. Datawrapper.
\url{https://www.datawrapper.de/blog/colorblindness-part2}
}
\keyword{datasets}
