% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting-text.R
\name{apa_words}
\alias{apa_words}
\title{Write APA Words}
\usage{
apa_words(x, ordinal = FALSE)
}
\arguments{
\item{x}{The number to be printed}

\item{ordinal}{Do you want the ordinal numbering (e.g., 1st, 6th, etc.)}
}
\value{
A character string
}
\description{
Confused about whether a number should be written out ("five") or use
numerals ("5")? Use this function! Most useful for R Markdown in-text
writing.
}
\examples{
apa_words(5)
apa_words(16)
apa_words(6, ordinal = TRUE)
}
