% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap_words.R
\name{cap_words}
\alias{cap_words}
\title{Capitalization function}
\usage{
cap_words(x)
}
\arguments{
\item{x}{a character string}
}
\description{
This function allows you to capitalize the first letter of every word in a character string.
}
\examples{
cap_words("how are you?")
}
\keyword{capitalization}
