% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gen.R
\name{clear_cache_ratlas}
\alias{clear_cache_ratlas}
\title{Clear memoise cache}
\usage{
clear_cache_ratlas()
}
\description{
Clear memoise cache
}
\details{
Clear memoise cache generated by \code{mem_get()}, a memoise function around
\code{\link[httr:GET]{httr::GET()}} used in \code{get_gen} in \code{get_singleton}
}
