% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors-ramps.R
\name{color_ramp}
\alias{color_ramp}
\alias{ramp_dlm}
\alias{ramp_okabeito}
\title{Custom color ramps}
\usage{
ramp_dlm(output, end = "#FFFFFF")

ramp_okabeito(output, end = "#FFFFFF", index = 1)
}
\arguments{
\item{output}{Colors to pull from the color ramp. Numbers range from 0-1,
which is a normalized sliding scale of the color ramp.}

\item{end}{The end color that the base color should fade into.}

\item{index}{For \code{\link[=ramp_okabeito]{ramp_okabeito()}}, the element of the palette to use.}
}
\value{
A character vector of hex color codes.
}
\description{
These color ramp functions create color scales that can be used for making
ggplot2 plots and gt tables.
}
\details{
\code{\link[=ramp_dlm]{ramp_dlm()}} is a pre-made color ramp based on the DLM blue.

\code{\link[=ramp_okabeito]{ramp_okabeito()}} is a pre-made color ramp based on the Okabe Ito palette.
}
\examples{
ramp_dlm(seq(0.1, 1), by = 0.2)
ramp_okabeito(seq(0.2, 1, length.out = 5), index = 2)
}
