% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read_table.R
\name{db_read_table}
\alias{db_read_table}
\title{Generic function to directly access data in Atlas database}
\usage{
db_read_table(
  table_name,
  schema = "public",
  output_geometry = FALSE,
  output_flatten = TRUE,
  limit = NULL,
  select = NULL,
  ...,
  .host = ATLAS_API_V4_HOST(),
  .token = ATLAS_API_TOKEN(),
  .max_active = 4,
  .page_limit = 10000,
  .header = list()
)
}
\arguments{
\item{table_name}{\code{character}. Name of the table to be accessed.}

\item{schema}{\code{character} default 'public'. Schema from the database where
the table is located. Can be either \code{public}, \code{api} or \code{atlas_api}.}

\item{output_geometry}{Optional. \code{logical} default \code{FALSE}. If \code{TRUE},
returns an \code{sf} object using the \code{geometry} column from the table.}

\item{output_flatten}{\code{logical} default \code{TRUE}. If \code{TRUE}, returns a
\code{data.frame} object with nested objects flattened.}

\item{limit}{Optional. \code{integer} default \code{NULL}. Maximum number of rows to
return. If \code{NULL}, all rows are returned. From the PostgREST API syntax.}

\item{select}{Optional. \code{character} default \code{NULL}. List of columns to
return. All columns are returned if \code{NULL}. From the PostgREST API syntax.}

\item{...}{Additional parameters to be passed as query to the API table_name.}

\item{.host}{Optional. \code{character} Atlas API host url.}

\item{.token}{Optional. \code{character} Bearer token providing access to the web api.}

\item{.max_active}{Optional. \code{integer} default \code{4}. Number of active connections
that are alive at all time to process the request.}

\item{.page_limit}{Optional. \code{integer} default \code{10000}. Maximum number of
rows to download per page. This determines how many requests (pages) the
download is split into.}

\item{.header}{\code{list} Additional headers to provide to the request.}
}
\value{
\code{tibble} or \code{sf} with rows associated with Atlas data object.
}
\description{
Return data objects stored in tables and views in the Atlas database.
}
\details{
This function is designed to interface with a web API deployed with
PostgREST.
}
