% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-manipulation.R
\name{df_append}
\alias{df_append}
\title{Append row and/or column summaries}
\usage{
df_append(df, row = TRUE, col = TRUE, .f = sum)
}
\arguments{
\item{df}{A data frame to append summaries to.}

\item{row}{logical indicating whether a summary row should be added (i.e., summarizing each column)}

\item{col}{logical indicating whether a summary column should be added (i.e., summarizing each row)}

\item{.f}{Function to use for calculating summaries}
}
\value{
A data frame with the summary row and/or column appended
}
\description{
Add row and/or column summaries (e.g., total counts) to a data frame.
}
\examples{
set.seed(9416)
df <- tibble::tibble(char = letters[1:5], x = rnorm(5), y = rnorm(5))
df_append(df, row = TRUE, col = TRUE, .f = sum)
df_append(df, row = FALSE, .f = mean)
}
