% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dlm_ll_info}
\alias{dlm_ll_info}
\title{DLM Linkage Level Information}
\format{
A data frame with 13 rows and 4 variables:
\itemize{
\item \code{subject}: The subject area for the linkage level
\item \code{linkage_level}: The linkage level name as it is stored in the database
\item \code{name}: The official linkage level name (public facing)
\item \code{value}: The value used for scoring (i.e., the linkage level order, within
subject)
}
}
\usage{
dlm_ll_info
}
\description{
A dataset contain information for how linkage levels are stored in the
database, their official names, and their values for scoring.
}
\keyword{datasets}
