% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dlm_ll_info}
\alias{dlm_ll_info}
\alias{dlm_labels}
\title{Dynamic Learning Maps naming shortcuts}
\format{
\code{dlm_ll_info} is a data frame with 13 rows and 4 variables:
\itemize{
\item \code{subject}: The subject area for the linkage level
\item \code{linkage_level}: The linkage level name as it is stored in the database
\item \code{name}: The official linkage level name (public facing)
\item \code{value}: The value used for scoring (i.e., the linkage level order, within
subject)
}

\code{dlm_labels} is a list with the following elements:
\itemize{
\item \code{linkage_levels}: The linkage level names for ELA, mathematics, and
science, in increasing complexity.
\item \code{grades}: The possible grade levels, grade bands, and end-of-year courses,
in increasing order.
\item \verb{complexity bands}: The four complexity band names for ELA, mathematics,
science, writing, and expressive communication, in increasing order.
\item \code{performance_levels}: The common performance level names (e.g.,
"Approaching"), in increasing order.
\item \code{performance_levels_full}: The official performance level names (e.g.,
"Approaching the Target"), in increasing order.
}
}
\usage{
dlm_ll_info

dlm_labels
}
\description{
Commonly used names and labels for the Dynamic Learning
Maps\ifelse{html}{\out{<sup>&reg;</sup>}}{\eqn{^\circledR}}
(DLM\ifelse{html}{\out{<sup>&reg;</sup>}}{\eqn{^\circledR}}) assessments.
}
\keyword{datasets}
