% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting-text.R
\name{fmt_italic}
\alias{fmt_italic}
\title{Format a string with italics for HTML or LaTeX output}
\usage{
fmt_italic(string, indicator = "*", html = TRUE)
}
\arguments{
\item{string}{The character vector to add italics to.}

\item{indicator}{The indicator for which words should be italicized.}

\item{html}{Logical for whether the output should be HTML. If \code{FALSE}, LaTeX
output is provided.}
}
\value{
A character vector.
}
\description{
Format a string with italics for HTML or LaTeX output
}
\examples{
fmt_italic("Make *this* italic.", html = TRUE)
fmt_italic("Make *this* italic.", html = FALSE)
}
\concept{HTML}
\concept{LaTeX}
