% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting-dataframe.R
\name{fmt_table}
\alias{fmt_table}
\title{Center and Decimal Align Tables}
\usage{
fmt_table(
  df,
  dec_dig = 1,
  prop_dig = 3,
  corr_dig = 3,
  output = NULL,
  keep_boundary = FALSE,
  ...
)
}
\arguments{
\item{df}{A data frame or tibble to be formatted for printing in output.}

\item{dec_dig}{The number of decimal places to include for numbers, e.g.,
\code{dec_dig = 1} for 16.5.}

\item{prop_dig}{The number of decimal places to include for numbers bounded
between [0,1], e.g., \code{prop_dig = 2} for .35.}

\item{corr_dig}{The number of decimal places to include for numbers bounded
between [-1,1], e.g., \code{corr_dig = 3} for .205.}

\item{output}{The output format of the table. One of "latex" or "html".
Automatically pulled from document output type if not specified.}

\item{keep_boundary}{Whether to preserve true values of boundaries (i.e.,
\code{min_value} and \code{max_value}). For example if \code{digits = 3}, \code{min_value = 0},
and \code{keep_boundary = FALSE}, a value exactly equal to 0 will become
"<.001". If \code{keep_boundary = TRUE}, then a value of 0 will remain "0.000",
and other small values (e.g., 0.000001) will continue to be replaced with
"<.001".}

\item{...}{Additional arguments passed to \code{\link[wjake:fmt_digits]{wjake::fmt_digits()}}.}
}
\value{
A tibble with the same rows and columns as \code{df}, with numbers
formatted consistently and padded for alignment when printed.
}
\description{
Automatic formatting for tables that should "just work" for most use cases.
For more fine-grained control, see \code{\link[wjake:fmt_digits]{wjake::fmt_digits()}} and
\link{padding}.
}
\examples{
pcts <- data.frame(n = 0:5, p = 0.5 * (0:5))
fmt_table(pcts)
}
\seealso{
Other formatters:
\code{\link{padding}}
}
\concept{formatters}
