% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cap.R
\name{format_cap}
\alias{format_cap}
\title{Word capitalization function}
\usage{
format_cap(x, allwords = FALSE)
}
\arguments{
\item{x}{A character string}

\item{allwords}{Defaults to FALSE. If TRUE, the first letter of every word in the string will be capitalized.}
}
\description{
This function allows you to capitalize first letters of the first word or every word in a string.
}
\examples{
format_cap("hello world!")
format_cap("hello world!", allwords = TRUE)
}
\keyword{capitalization}
\keyword{upper}
