% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{formatting}
\alias{formatting}
\alias{fmt_digits}
\alias{fmt_leading_zero}
\alias{fmt_minus}
\alias{fmt_replace_na}
\alias{fmt_corr}
\alias{fmt_prop}
\title{Text and Number Formatting}
\usage{
fmt_digits(x, digits = 3)

fmt_leading_zero(x)

fmt_minus(x)

fmt_replace_na(x, replacement = "&mdash;")

fmt_corr(x, digits)

fmt_prop(x, digits, fmt_small = TRUE)
}
\arguments{
\item{x}{Number or number string to be formatted}

\item{digits}{Number of decimal places to retain}

\item{replacement}{The value to use when replacing missing values}

\item{fmt_small}{Indicator for replacing zero with \code{<} (e.g., \code{.000} becomes
\verb{< .001}). Default is \code{TRUE}.}
}
\description{
These formatting functions are used to format numerical values in a
consistent manner. This is useful for printing numbers inline with text, as
well as for formatting tables. Many of the included formatting functions were
adapted from TJ Mahr's \href{https://github.com/tjmahr/printy}{printy} package.
}
\details{
\code{fmt_digits()} is a wrapper for \code{\link[base:sprintf]{base::sprintf()}}. Prints a number with
\code{digits} number of decimal places, without losing trailing zeros, as happens
with \code{\link[base:round]{base::round()}}.

\code{fmt_leading_zero()} removes the leading zero for decimal values.

\code{fmt_minus()} replaces hyphens with the HTML minus sign (\verb{&minus;}).

\code{fmt_replace_na()} replaces \code{NA} values with a specified replacement. This is
useful for formatting tables, when blanks are not desired. The default
behavior is to replace missing values with an em-dash (\verb{&mdash;}).

Two additional formatters are provided to format numbers according to the
American Psychological Association (APA) style guide. The 7th edition of the
APA style guide specifies that numbers bounded between [-1, 1] should not
include the leading zero (section 6.36; APA, 2020). This is the case for many
types of numbers commonly used by ATLAS including correlations, proportions,
probabilities, and \emph{p}-values. The \code{fmt_corr()} function is used to format
values bounded between [-1, 1]. Digits are first rounded to the specified
number of digits using \code{fmt_digits()}, and then leading zeros are removed
using \code{fmt_leading_zero()} and negative signs are replaced with
\code{fmt_minus()}. The \code{fmt_prop} is very similar, but is intended for values
between [0, 1]. This function also wraps \code{fmt_digits()} and
\code{fmt_leading_zero()}. However, \code{fmt_prop()} also replaces small values to
avoid values of \code{0} (e.g., \code{.00} is replaced with \verb{< .01}).
}
\examples{
test_cor <- cor(mtcars[, 1:4])
as.character(round(test_cor[1:4, 3], 2))
fmt_digits(test_cor[1:4, 3], 2)

fmt_digits(test_cor[1:4, 3], 2) \%>\%
  fmt_leading_zero()

fmt_digits(test_cor[1:4, 3], 2) \%>\%
  fmt_minus()

fmt_digits(c(test_cor[1:4, 3], NA_real_), 2) \%>\%
  fmt_replace_na(replacement = "&mdash;")

fmt_corr(test_cor[1:4, 3], 2)

fmt_prop(c(0.001, 0.035, 0.683), digits = 2)

}
\references{
American Psychological Association. (2020). \emph{Publication manual
of the American Psychological Association} (7th ed.).
\url{https://doi.org/10.1037/0000165-000}
}
\concept{formatters}
