% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bird_presence_absence.R
\name{get_bird_presence_absence}
\alias{get_bird_presence_absence}
\title{Get bird presence absence}
\usage{
get_bird_presence_absence(taxa_name, .cores = 4, .page_limit = 4e+05)
}
\arguments{
\item{.cores}{Optional. \code{integer} default \code{4}. Number of cores used to
parallelize and improve rapidity}

\item{id_taxa}{Optional. \code{integer} scalar or vector. Returns a dataframe
for the observations related to the \code{id_taxa}}
}
\value{
\code{tibble} with rows associated with Atlas observations
}
\description{
The function downloads and returns a dataframe containing presence and
absence (values TRUE or FALSE) with a single record per row for the bird
species specified by \code{id_taxa}.
}
\details{
The absence are infered from all sampling points related to birds observed
on Quebec territory whose observations were obtained from datasets
defined as \code{exhaustive}.
}
\examples{
# Returns all taxa filtered by the column `id_taxa` values
pres_abs <- get_bird_presence_absence(id_taxa = 188)

}
