% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.R
\name{get_group}
\alias{get_group}
\title{get_group
Find a match between classification of the taxon and the group object}
\usage{
get_group(classification)
}
\description{
get_group
Find a match between classification of the taxon and the group object
}
