% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mtl_bird_observations.R
\name{get_mtl_bird_observations}
\alias{get_mtl_bird_observations}
\title{Get mtl bird observations}
\usage{
get_mtl_bird_observations(id = NULL, year = NULL, id_taxa = NULL, ...)
}
\arguments{
\item{id}{Optional. \code{integer} scalar or vector. Returns a dataframe for the
observation with the specified id}

\item{year}{Optional. \code{integer} scalar or vector. Returns a dataframe for
the observations related to the \code{id_taxa}}

\item{id_taxa}{Optional. \code{integer} scalar or vector. Returns a dataframe
for the taxa_obs record related to the value. \code{id_taxa} is translated to
\code{id_taxa_obs}}

\item{...}{Optional. scalar or vector. Returns a dataframe filtered by the
atlas \code{observations} table columns specified as parameter
(ie. \code{id_datasets}, \code{id_variables}, \code{month_obs}, \code{day_obs})}

\item{.cores}{Optional. \code{integer} default \code{4}. Number of cores used to
parallelize and improve rapidity}
}
\value{
\code{tibble} with rows associated with Atlas observations
}
\description{
The function downloads and returns a list or dataframe containing
observations with a single record per row. The function returns a dataframe
containing all records if no parameters are specified (STRONGLY DISCOURAGED
DUE TO LARGE AMOUNT OF OBSERVATIONS CONTAINED WITHIN ATLAS).
The function filters returned records by attributes corresponding to atlas
table columns specified as parameters (ie. \code{id}, \code{year_obs}, \code{id_taxa},
\code{id_datasets}, \code{id_variables}, \code{etc}) with accepted values either being
scalar or vector for single or multiple records.
}
\details{
The absence are infered from all sampling points related to birds observed
on Montreal territory whose observations were obtained from datasets
defined as \code{exhaustive}.
}
