% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mtl_bird_presence_absence.R
\name{get_mtl_bird_presence_absence}
\alias{get_mtl_bird_presence_absence}
\title{Get Montreal bird presence absence}
\usage{
get_mtl_bird_presence_absence(taxa_name, .cores = 4)
}
\arguments{
\item{.cores}{Optional. \code{integer} default \code{4}. Number of cores used to
parallelize and improve rapidity}

\item{taxa_name.}{\code{str} Returns a dataframe
for the observations related to the to the\code{scientific_name} corresponding to
any rank.}
}
\value{
\code{tibble} with rows associated with Atlas observations
}
\description{
The function downloads and returns a dataframe containing presence and
absence (values TRUE or FALSE) with a single record per row for the bird
species specified by \code{id_taxa}.
}
\details{
The absence are infered from all sampling points related to birds observed
on Montreal territory whose observations were obtained from datasets
defined as \code{exhaustive}.
}
\examples{
# Returns all presence absence for the blue jay `Cyanocitta cristata`
pres_abs <- get_mtl_bird_presence_absence(taxa_name = "Cyanocitta cristata")

}
