% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_data.R
\name{get_table_data}
\alias{get_table_data}
\title{Generic function to access data from tables and views in Atlas database}
\usage{
get_table_data(
  table_name,
  schema = "public",
  output_geometry = FALSE,
  output_flatten = TRUE,
  limit = NULL,
  select = NULL,
  ...,
  .token = NULL,
  .cores = 4,
  .n_pages = NULL,
  .page_limit = 10000
)
}
\arguments{
\item{table_name}{\code{character}. Name of the table to be accessed.}

\item{schema}{\code{character} default 'public'. Schema from the database where
the table is located.
Can be either \code{public}, \code{api}, \code{public_api}, or \code{atlas_api}.}

\item{output_geometry}{Optional. \code{logical} default \code{FALSE}. If \code{TRUE},
returns an \code{sf} object using the \code{geometry} column from the table.}

\item{output_flatten}{\code{logical} default \code{TRUE}. If \code{TRUE}, returns a
\code{data.frame} object with nested objects flattened.}

\item{limit}{Optional. \code{integer} default \code{NULL}. Maximum number of rows to
return. If \code{NULL}, all rows are returned. From the PostgREST API syntax.}

\item{select}{Optional. \code{character} default \code{NULL}. List of columns to
return. All columns are returned if \code{NULL}. From the PostgREST API syntax.}

\item{...}{Additional parameters to be passed as query to the API table_name.}

\item{.token}{Optional. \code{character} Bearer token providing access to the web
API. If \code{NULL}, the function will attempt to read the token from the
\code{ATLAS_API_TOKEN} environment variable.}

\item{.cores}{Optional. \code{integer} default \code{4}. Number of cores to use for
parallel processing. If \code{1}, no parallel processing is used. If \code{NULL}, the
function will automatically determine the number of cores to use based on
the number of available cores on the system.}

\item{.n_pages}{Optional. \code{integer} default \code{NULL}. Number of pages to
download. ' If \code{NULL}, the number of pages is estimated from the number of
rows in the table.}

\item{.page_limit}{Optional. \code{integer} default \code{500000}. Maximum number of
rows to download per page. This parameter is used to estimate the number of
pages to download if \code{.n_pages} is \code{NULL}.}
}
\value{
\code{tibble} or \code{sf} with rows associated with Atlas data object
}
\description{
Return data objects stored in tables and views in the Atlas database.
}
\details{
This function is designed to interface with a web API deployed with
PostgREST.
}
