% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{get_timeseries}
\alias{get_timeseries}
\title{Get timeseries}
\usage{
get_timeseries(id = NULL, id_taxa_obs = NULL, ...)
}
\arguments{
\item{id}{Returns a dataframe for the population with the specified id}

\item{id_taxa_obs}{Returns a dataframe containing time series with species}

\item{...}{Optional. scalar or vector. Returns a dataframe filtered by the
atlas \code{observations} table columns specified as parameter}
}
\value{
A list or dataframe with rows corresponding to individual population.
}
\description{
The function downloads and returns a spatial feature (sf) dataframe
containing time-series, for Quebec, where each row corresponds to an
individual population.
}
