% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_mtl_bird_taxa.R
\name{list_mtl_bird_taxa}
\alias{list_mtl_bird_taxa}
\title{List Montreal bird taxa}
\usage{
list_mtl_bird_taxa(rank = NULL, source_name = NULL, valid = NULL, ...)
}
\arguments{
\item{rank}{Optional. \code{character} Filter by rank, such as \code{family}, \code{species}
, \code{class}, etc.}

\item{source_name}{Optional. \code{character} Filter by taxonomic source.
Accepted values are \code{ITIS}, \verb{Catalogue of Life} and \verb{GBIF Backbone Taxonomy}}

\item{valid}{Optional. \code{logical}}

\item{...}{Optional. scalar or vector. Returns a dataframe filtered by the
atlas \code{taxa_ref} table columns specified as parameter}
}
\value{
\code{tibble} with rows associated with Atlas \code{taxa_ref} records
}
\description{
The function downloads and returns a dataframe containing presence and
absence (values TRUE or FALSE) with a single record per row for the bird
species specified by \code{id_taxa}.
}
\details{
The absence are infered from all sampling points related to birds observed
on Montreal territory whose observations were obtained from datasets
defined as \code{exhaustive}.
}
\examples{
# Returns all bird species of different sources
pres_abs <- get_mtl_bird_presence_absence(rank = "species")

}
