% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{measr_pdf}
\alias{measr_pdf}
\title{Create an R Markdown PDF measr Report}
\usage{
measr_pdf(...)
}
\arguments{
\item{...}{Arguments to be passed to \verb{[bookdown::pdf_document2]}}
}
\value{
A modified \code{pdf_document2} with the standard tech report formatting.
}
\description{
This is a function called in the output of the yaml of the Rmd file to
specify using the standard measr report document formatting.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
output: ratlas::measr_pdf
\dontshow{\}) # examplesIf}
}
