% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-atlas-ms.R
\docType{data}
\name{font_ms}
\alias{font_ms}
\title{Montserrat font name R variable aliases}
\format{length 1 character vector}
\usage{
font_ms
}
\description{
\code{font_ms} == "\code{Montserrat}"
}
\keyword{datasets}
