% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num2word.R
\name{num2word}
\alias{num2word}
\title{Number converting function}
\usage{
num2word(x)
}
\arguments{
\item{x}{Ideally an integer from 0 to 9. If x is already a written number or any integer not in 0-9, then this function will return x.}
}
\description{
This function allows you to convert an integer from 0 to 9 to a text representation of the integer. For example, this function will convert the number 4, to the word "four".
}
\examples{
num2word(4)
num2word("five")
num2word(-99)
}
\keyword{convert}
\keyword{integer}
\keyword{number}
\keyword{text}
