% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_format.R
\name{num_format}
\alias{num_format}
\title{Formatting and rounding numbers function}
\usage{
num_format(x, digits = 3)
}
\arguments{
\item{x}{A numeric value}

\item{digits}{Number of decimal places. Defaults to 3.}
}
\description{
This function allows you to format or round a number to a specified number of decimal places.
}
\examples{
num_format(2)
num_format(340.5892)
num_format(340.5892, digits = 2)
}
\keyword{format}
\keyword{numbers}
\keyword{round}
