% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-manipulation.R
\name{only_if}
\alias{only_if}
\title{Only If}
\usage{
only_if(condition)
}
\arguments{
\item{condition}{Logical condition to be evaluated}
}
\description{
Adverb for conditionally skipping steps in a piped workflow.
}
\examples{
d <- tibble::as_tibble(mtcars)
d \%>\%
  only_if(TRUE)(dplyr::filter)(.data$mpg > 25)

d \%>\%
  only_if(FALSE)(dplyr::filter)(.data$mpg > 25)
}
\author{
David Robinson, https://twitter.com/drob/status/785880369073500161
}
