% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_num.R
\name{order_num}
\alias{order_num}
\title{Number ordering function}
\usage{
order_num(x)
}
\arguments{
\item{x}{Ideally an integer from 1 to 10. If x is already a written ordinal number or any integer not in [1, 10], then this function will return x.}
}
\description{
This function allows you to convert an integer from 1 to 10 to its corresponding ordinal number. For example, this function will convert the number 4 to its ordinal number "fourth".
}
\examples{
order_num(4)
order_num("fifth")
order_num(-99)
}
\keyword{convert}
\keyword{integer}
\keyword{number}
\keyword{ordinal}
\keyword{text}
