% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{report_docx}
\alias{report_docx}
\alias{report_rdocx}
\alias{report_pdf}
\alias{report_gitbook}
\title{Create an R Markdown Word Document Topic Guide}
\usage{
report_docx(type = "topicguide", ...)

report_rdocx(type = "topicguide", ...)

report_pdf(type = "topicguide", ...)

report_gitbook(type = "techreport", ...)
}
\arguments{
\item{type}{The type of report. Should be one of "topicguide", "techreport",
or "measr". The following outputs are supported for each type:
\itemize{
\item topicguide: docx, rdocx, pdf
\item techreport: pdf, gitbook
\item measr: pdf
}}

\item{...}{Arguments to be passed to relevant {bookdown} or {officedown}
output functions:
\itemize{
\item \code{\link[=report_docx]{report_docx()}} passes to \code{\link[bookdown:html_document2]{bookdown::word_document2()}}
\item \code{\link[=report_rdocx]{report_rdocx()}} passes to \code{\link[officedown:rdocx_document]{officedown::rdocx_document()}}
\item \code{\link[=report_pdf]{report_pdf()}} passes to \code{\link[bookdown:html_document2]{bookdown::pdf_document2()}}
\item \code{\link[=report_gitbook]{report_gitbook()}} passes to \code{\link[bookdown:gitbook]{bookdown::gitbook()}}
}}
}
\value{
A modified output format with the appropriate template and default
options defined.
}
\description{
This is a function called in the output of the yaml of the Rmd file to
specify using the standard DLM topic guide word document formatting.
}
\examples{
\dontrun{
output: ratlas::report_docx
}
}
