% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-atlas-pl.R
\docType{data}
\name{font_pl}
\alias{font_pl}
\title{Palatino font name R variable aliases}
\format{
length 1 character vector
}
\usage{
font_pl
}
\description{
\code{font_pl} == "\code{Palatino}"
}
\keyword{datasets}
