% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{palette_atlas}
\alias{palette_atlas}
\alias{palette_atlas_black}
\title{Official color palette for ATLAS}
\format{
An object of class \code{character} of length 6.

An object of class \code{character} of length 6.
}
\usage{
palette_atlas

palette_atlas_black
}
\description{
Official brand colors for Accessible Teaching, Learning, and Assessment
Systems.
}
\keyword{datasets}
