% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{palette_lcrost}
\alias{palette_lcrost}
\alias{palette_lcrost_black}
\title{Color palette proposed by Lisa Charlotte Rost}
\format{
An object of class \code{character} of length 8.

An object of class \code{character} of length 8.
}
\usage{
palette_lcrost

palette_lcrost_black
}
\description{
A colorblind friendly palette taken from the article \href{https://www.datawrapper.de/blog/colorblindness-part2}{What to consider when visualizing data for colorblind readers}
}
\keyword{datasets}
