% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{palette_okabe_ito}
\alias{palette_okabe_ito}
\alias{palette_okabe_ito_black}
\title{Color palette proposed by Okabe and Ito}
\format{An object of class \code{character} of length 8.}
\usage{
palette_okabe_ito

palette_okabe_ito_black
}
\description{
Two color palettes taken from the article "Color Universal Design" by Okabe
and Ito, http://jfly.iam.u-tokyo.ac.jp/color/. The variant
\code{palette_okabe_ito} contains a gray color, while \code{palette_okabe_ito_black}
contains black instead.
}
\keyword{datasets}
