% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{palette_okabeito}
\alias{palette_okabeito}
\alias{palette_okabeito_black}
\title{Color palette proposed by Okabe and Ito}
\format{
An object of class \code{character} of length 8.

An object of class \code{character} of length 8.
}
\usage{
palette_okabeito

palette_okabeito_black
}
\description{
Two color palettes taken from the article "Color Universal Design" by Okabe
and Ito, http://jfly.iam.u-tokyo.ac.jp/color/. The variant
\code{palette_okabeito} contains a gray color, while \code{palette_okabeito_black}
contains black instead.
}
\keyword{datasets}
