% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_dataset.R
\name{post_dataset}
\alias{post_dataset}
\title{Function to insert datasets in Atlas.
Formatting should follow https://eml.ecoinformatics.org/schema/}
\usage{
post_dataset(
  title,
  creator,
  id = NULL,
  pub_date = NULL,
  abstract = NULL,
  keyword_set = NULL,
  distribution = NULL,
  alternate_identifier = NULL,
  licensed = NULL,
  associated_party = NULL,
  contact,
  metadata_provider = NULL,
  additional_info = NULL,
  language = NULL,
  publisher = NULL,
  coverage = NULL,
  methods = NULL,
  additional_metadata = NULL,
  data_type,
  source_eml_url = NULL,
  source_alias,
  shareable_data = FALSE,
  schema = "api"
)
}
\arguments{
\item{title}{\code{character}. Name of the dataset.}

\item{creator}{\code{list}. List of responsibleParty objects created with EML::set_responsibleParty().}

\item{id}{Optional. \code{character}. UUID of the dataset. If not provided, a new UUID will be generated.}

\item{pub_date}{Optional. \code{character}. Publication date of the dataset in "YYYY-MM-DD" format.}

\item{abstract}{Optional. \code{character}. Abstract of the dataset.}

\item{keyword_set}{Optional. \code{list}. List of keywordSet objects created manually.}

\item{distribution}{Optional. \code{list}. List of list distribution objects created manually.}

\item{alternate_identifier}{Optional. \code{list}. List of alternateIdentifier \code{character}. DOI is declared here.}

\item{licensed}{Optional. \code{list}. List of license informations created manually.}

\item{associated_party}{Optional. \code{list}. List of responsibleParty objects created manually.}

\item{contact}{\code{list}. List of responsibleParty objects created with EML::set_responsibleParty().}

\item{metadata_provider}{Optional. \code{list}. List of responsibleParty objects created with EML::set_responsibleParty().}

\item{additional_info}{Optional. \code{list}. List of additional information about the dataset created with EML::set_TextType() to a markdown tag.}

\item{language}{Optional. \code{character}. Language of the dataset.}

\item{publisher}{Optional. \code{list}. List of responsibleParty objects created with EML::set_responsibleParty().}

\item{coverage}{Optional. \code{list}. List of coverage objects created with EML::set_coverage().}

\item{methods}{Optional. \code{jsonb}. Methods section created with EML::set_methods() or manually.}

\item{additional_metadata}{Optional. \code{character}. Citation string, nested in metadata, citation.

Non-EML parameters:}

\item{data_type}{\code{character}. Type of data, example: occurrence, time series, etc.}

\item{source_eml_url}{Optional. \code{character}. URL of the source EML.}

\item{source_alias}{\code{character}. Internal alias to recognize the dataset.}

\item{shareable_data}{\code{boolean}. Default \code{FALSE}, if \code{TRUE} it means we have the license to share the data.}

\item{schema}{obligatory api, as function resides in this schema on the backend.}
}
\value{
newly generated dataset UUID.
}
\description{
This function is designed to interface with a web API deployed with PostgREST
Function is not exported, as it is meant for internal use only.
It can be called like so: ratlas:::post_dataset()
}
\details{
EML parameters:
}
\keyword{internal}
