% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_gen.R
\name{post_gen}
\alias{post_gen}
\title{Generic function to post data into Atlas databases}
\usage{
post_gen(
  endpoint,
  data,
  ...,
  .schema = "public",
  .page_limit = 50000,
  .token = ATLAS_API_TOKEN(),
  .cores = 4
)
}
\arguments{
\item{endpoint}{\code{character}. Name of the atlas data object destination.
Corresponds to the name of a within Atlas Postgresql database, stored within
schema \code{public} ou \code{api}}

\item{data}{\code{list} or \code{data.frame}. Either a single record as a \code{list} with
names as attributes or multiple records as a \code{data.frame} with attributes as
columns with attibutes corresponding to the destination column tables.}

\item{...}{\code{character} or \code{numeric} scalar or vector.
Additional parameters to provide to the request. May correspond to postgrest
http request parameters and syntax.}

\item{.schema}{\code{character} Schema from the database where is located the data
object is located. Accept either values \code{api} or \code{public} (default)}

\item{.page_limit}{\code{integer} Count of objects returned through pagination}

\item{.token}{\code{character} Bearer token providing access to the web api}

\item{.cores}{\code{integer} default \code{4}. Number of cores used to parallelize and
improve rapidity}
}
\value{

}
\description{
Data records attributes (list names or columns) must corresponds to the
table columns it is sent to.
(ie. taxa, observations, datasets, etc).
}
\details{
This function is designed to interface with a web API deployed with PostgREST
}
