% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting-plots.R
\name{scale_x_comma}
\alias{scale_x_comma}
\alias{scale_y_comma}
\alias{scale_x_percent}
\alias{scale_y_percent}
\title{X & Y scales for percent and comma label formats}
\usage{
scale_x_comma(...)

scale_y_comma(...)

scale_x_percent(accuracy = NULL, ...)

scale_y_percent(accuracy = NULL, ...)
}
\arguments{
\item{...}{Arguments passed to
\code{\link[ggplot2:scale_continuous]{ggplot2::scale_(x|y)_continuous}}.}

\item{accuracy}{A number to round to. Use (e.g.) \code{0.01} to show 2 decimal
places of precision. If \code{NULL}, the default, uses a heuristic that should
ensure breaks have the minimum number of digits needed to show the
difference between adjacent values.

Applied to rescaled data.}
}
\value{
A scale for {ggplot2}.
}
\description{
The \verb{_comma} scales set a comma format (e.g., "1,000" vs. "1000") for axis
labels. The \verb{_percent} scales round numbers and add a percentage sign to the
label.
}
\examples{
dat <- tibble::tibble(count = seq(0, 10000, by = 1000),
                      percent = seq(0, 1, by = 0.1))

ggplot(dat, aes(x = count, y = percent)) +
  geom_point() +
  scale_x_comma() +
  scale_y_percent()
}
