% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rat_apa_decimal.R
\name{rat_apa_decimal}
\alias{rat_apa_decimal}
\title{Format decimals in APA style}
\usage{
rat_apa_decimal(x, digits = 3)
}
\arguments{
\item{x}{A numeric value}

\item{digits}{Number of decimal places. Defaults to 3.}
}
\value{
A character string of the formatted and rounded number
}
\description{
A function for printing decimals in APA style. If the number cannot exceed -1
to 1, then the leading zero should not be printed. This function takes a
decimal between -1 and 1, rounds to the specified number of places, and then
formats without the leading 0.
}
