% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rat_format_cap.R
\name{rat_format_cap}
\alias{rat_format_cap}
\title{Word capitalization function}
\usage{
rat_format_cap(x, allwords = FALSE)
}
\arguments{
\item{x}{Must be a character string}

\item{allwords}{Defaults to FALSE. If TRUE, the first letter of every word in the string will be capitalized.}
}
\description{
This function allows you to capitalize first letter of the first word or every word in a character string.
}
\examples{
rat_format_cap("hello world!")
rat_format_cap("hello world!", allwords = TRUE)
}
\keyword{capitalization}
\keyword{upper}
