% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-bin.R
\docType{data}
\name{ratlas-ggproto}
\alias{ratlas-ggproto}
\alias{StatBin2}
\title{Base ggproto classes for ratlas}
\description{
Base ggproto classes for ratlas
}
\seealso{
\link{ggproto}
}
\keyword{datasets}
\keyword{internal}
