% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratlas-package.R
\docType{package}
\name{ratlas-package}
\alias{ratlas}
\alias{ratlas-package}
\title{ratlas: ATLAS Formatting Functions and Templates}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides templates, formatting tools, and 'ggplot2' themes tailored for the Accessible Teaching, Learning, and Assessment Systems (ATLAS) organization. These templates facilitate the creation of topic guides and technical reports, while the formatting functions enable users to customize numbers and tables to meet specific requirements. Additionally, the themes ensure a uniform visual style across graphics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ratlas.netlify.app}
  \item \url{https://github.com/atlas-aai/ratlas}
  \item Report bugs at \url{https://github.com/atlas-aai/ratlas/issues}
}

}
\author{
\strong{Maintainer}: W. Jake Thompson \email{wjakethompson@gmail.com} (\href{https://orcid.org/0000-0001-7339-0300}{ORCID})

Authors:
\itemize{
  \item Noelle Pablo \email{noellepablo@gmail.com}
  \item Jeffrey Hoover \email{jeffrey.c.hoover@gmail.com} (\href{https://orcid.org/0000-0002-0276-0308}{ORCID})
}

Other contributors:
\itemize{
  \item University of Kansas [copyright holder, funder]
}

}
\keyword{internal}
