% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{margin}
\alias{palette_okabeito}
\alias{scale_color_okabeito}
\alias{scale_colour_okabeito}
\alias{scale_fill_okabeito}
\alias{scale_x_comma}
\alias{scale_x_percent}
\alias{scale_y_comma}
\alias{scale_y_percent}
\alias{apa_words}
\alias{gt_theme_apa}
\alias{fmt_corr}
\alias{fmt_count}
\alias{fmt_digits}
\alias{fmt_pct}
\alias{fmt_prop}
\alias{fmt_prop_pct}
\alias{round_to}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2:margin]{margin()}}}

  \item{wjake}{\code{\link[wjake:apa_words]{apa_words()}}, \code{\link[wjake:fmt_corr]{fmt_corr()}}, \code{\link[wjake:fmt_count]{fmt_count()}}, \code{\link[wjake:fmt_digits]{fmt_digits()}}, \code{\link[wjake:fmt_pct]{fmt_pct()}}, \code{\link[wjake:fmt_prop]{fmt_prop()}}, \code{\link[wjake:fmt_prop_pct]{fmt_prop_pct()}}, \code{\link[wjake:gt_theme_apa]{gt_theme_apa()}}, \code{\link[wjake:palette_okabeito]{palette_okabeito}}, \code{\link[wjake:round_to]{round_to()}}, \code{\link[wjake:scale_color_okabeito]{scale_color_okabeito()}}, \code{\link[wjake:scale_colour_okabeito]{scale_colour_okabeito()}}, \code{\link[wjake:scale_fill_okabeito]{scale_fill_okabeito()}}, \code{\link[wjake:scale_x_comma]{scale_x_comma()}}, \code{\link[wjake:scale_x_percent]{scale_x_percent()}}, \code{\link[wjake:scale_y_comma]{scale_y_comma()}}, \code{\link[wjake:scale_y_percent]{scale_y_percent()}}}
}}

