% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessible-color-scales.R
\name{scale_colour_accessible}
\alias{scale_colour_accessible}
\alias{scale_color_accessible}
\alias{scale_fill_accessible}
\alias{scale_accessible}
\title{Accessible color scales}
\usage{
scale_colour_accessible(aesthetics = "colour", ...)

scale_color_accessible(aesthetics = "colour", ...)

scale_fill_accessible(aesthetics = "fill", ...)

scale_accessible(
  aesthetics,
  option = c("okabeito", "datawrapper"),
  use_black = FALSE,
  order = 1:8,
  darken = 0,
  alpha = NA,
  ...
)
}
\arguments{
\item{aesthetics}{The names of the aesthetics that this scale works with.}

\item{...}{Additional arguments passed to \link[ggplot2:discrete_scale]{ggplot2::discrete_scale}.}

\item{option}{The color palette to use. One of "okabeito" or "datawrapper".}

\item{use_black}{If \code{TRUE}, scale includes black, otherwise includes gray.}

\item{order}{Numeric vector listing the order in which the colors should be
used. Default is 1:8.}

\item{darken}{Relative amount by which the scale should be darkened (for
positive values) or lightened (for negative values).}

\item{alpha}{Alpha transparency level of the color. Default is no
transparency.}
}
\value{
A color scale for use in plots created with {ggplot2}.
}
\description{
These functions provide {ggplot2} wrappers around the
\link[=accessible-palettes]{accessible color palettes} provided by {ratlas}.
}
\examples{
library(ggplot2)

ggplot(iris, aes(Sepal.Length, Sepal.Width, color = Species)) +
  geom_point() +
  scale_color_accessible(type = "datawrapper")

ggplot(iris, aes(Sepal.Length, fill = Species)) +
  geom_density(alpha = 0.7) +
  scale_fill_accessible(type = "okabeito", order = c(1, 3, 5))
}
\references{
Muth, L. C. (2020, June 23). \emph{What to consider when visualizing
data for colorblind readers}. Datawrapper.
\url{https://www.datawrapper.de/blog/colorblindness-part2}

Okabe M., & Ito, K. (2008). \emph{Color universal design (CUD): How to
make figures and presentations that are friendly to Colorblind people}.
Jfly. \url{https://jfly.uni-koeln.de/color/}
}
\seealso{
\code{\link[ggplot2:scale_viridis]{ggplot2::scale_colour_viridis_d()}}
}
