% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-theme.R
\name{set_theme}
\alias{set_theme}
\title{Set default ggplot2 theme}
\usage{
set_theme(font = "Arial Narrow")
}
\arguments{
\item{font}{One of "Arial Narrow" (default) or "Montserrat".}
}
\description{
Sets the default color schemes, fonts, and theme for ggplot2 plots. The
default color scheme for continuous variables is the
\href{https://cran.r-project.org/web/packages/viridis/vignettes/intro-to-viridis.html}{viridis}
color palette, and the default color scheme for discrete variables is the
\href{http://jfly.iam.u-tokyo.ac.jp/color/}{Okabe Ito} palette.
}
\examples{
set_theme("Arial Narrow")
set_theme("Montserrat")

}
