% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-atlas.R
\name{set_theme_ratlas}
\alias{set_theme_ratlas}
\title{Set default ggplot2 theme}
\usage{
set_theme_ratlas(
  font = "Archivo Narrow",
  discrete = "okabeito",
  continuous = "mako",
  ...
)
}
\arguments{
\item{font}{The base font family to be used in plots.}

\item{discrete}{Color palette for discrete colors. One of "okabeito"
(default), "atlas", or "ggplot2".}

\item{continuous}{Color palette for continuous scales. One of "magma",
"inferno", "plasma", "viridis" (default), or "cividis", or "ggplot2".}

\item{...}{Additional arguments to passed to \code{\link[=theme_atlas]{theme_atlas()}}.}
}
\value{
None. Called for side effects.
}
\description{
Sets the default color schemes, fonts, and theme for ggplot2 plots. The
default color scheme for continuous variables is the
\href{https://CRAN.R-project.org/package=viridis}{viridis}
color palette, and the default color scheme for discrete variables is the
\href{https://jfly.uni-koeln.de/color/}{Okabe Ito} palette.
}
\examples{
p1 <- ggplot(penguins, aes(x = bill_len, y = flipper_len)) +
  geom_point(aes(color = species), na.rm = TRUE)
p1

set_theme_ratlas(font = "mono")
p1

reset_theme_settings()
p1
}
