% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{techreport_gitbook}
\alias{techreport_gitbook}
\title{Create an R Markdown GitBook Tech Report}
\usage{
techreport_gitbook(...)
}
\arguments{
\item{...}{Arguments to be passed to \verb{[bookdown::gitbook]}}
}
\value{
A modified \code{gitbook} with the standard tech report formatting.
}
\description{
This is a function called in the output of the yaml of the Rmd file to
specify using the standard DLM tech report pdf document formatting.
}
\examples{
\dontrun{
output: ratlas::techreport_gitbook
}
}
