% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_guide.R
\name{topic_guide}
\alias{topic_guide}
\title{Topic Guide Formatting Function}
\usage{
topic_guide(...)
}
\arguments{
\item{...}{arguments to be passed to \code{rmarkdown::\link[rmarkdown]{pdf_document}}}
}
\description{
This is a function called in output in the YAML of the driver Rmd file
to specify using the ATLAS/DLM topic guide template.
}
\examples{
\dontrun{
 output: ratlas::topic_guide
}
}
\keyword{cats}
