% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{topicguide_docx}
\alias{topicguide_docx}
\title{Create an R Markdown Word Document Topic Guide}
\usage{
topicguide_docx(...)
}
\arguments{
\item{...}{Arguments to be passed to \verb{[bookdown::word_document2]}}
}
\value{
A modified \code{word_document2} with the standard topic guide formatting.
}
\description{
This is a function called in the output of the yaml of the Rmd file to
specify using the standard DLM topic guide word document formatting.
}
\examples{
\dontrun{
  output: ratlas::topicguide_docx
}
}
