% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-atlas.R
\name{update_geom_font_defaults}
\alias{update_geom_font_defaults}
\title{Update matching font defaults for text geoms}
\usage{
update_geom_font_defaults(
  family = "Arial Narrow",
  face = "plain",
  size = 3.5,
  color = "#2b2b2b"
)
}
\arguments{
\item{family, face, size, color}{font family name, face, size and color}
}
\value{
None. Called for side effects.
}
\description{
Updates \link[ggplot2:geom_text]{ggplot2::geom_label} and \link[ggplot2:geom_text]{ggplot2::geom_text} font defaults
}
\examples{
# updates font to Arial Narrow, size to 3.5, and color to #2b2b2b by default
update_geom_font_defaults()

}
