% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-atlas-pl.R
\name{update_geom_font_pl_defaults}
\alias{update_geom_font_pl_defaults}
\title{Update matching font defaults for text geoms}
\usage{
update_geom_font_pl_defaults(family = "Palatino Linotype",
  face = "plain", size = 3.5, color = "#2b2b2b")
}
\arguments{
\item{family, face, size, color}{font family name, face, size and color}
}
\description{
Updates \link[ggplot2:geom_label]{ggplot2::geom_label} and \link[ggplot2:geom_text]{ggplot2::geom_text} font defaults
}
