% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_vect.R
\name{connect_vect}
\alias{connect_vect}
\title{Vector concatenating function}
\usage{
connect_vect(vec, final = "and")
}
\arguments{
\item{vec}{a vector containing either words, numbers, or both.}

\item{final}{A character object that connects the last object in vec to the rest of the objects in vec. Defaults to "and".}
}
\description{
This function allows you to turn a vector of words or numbers into a character string.
The last object in the vector will be connected to the rest of the objects with an "and" by default.
}
\examples{
connect_vect(c(4, 5, 6, 7))
connect_vect(c("apples", "oranges", "pears"), final = "or")
connect_vect(c("chicken", 2))
}
\keyword{concatenate}
\keyword{integer}
\keyword{number}
\keyword{paste}
\keyword{text}
