% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_write_table.R
\name{db_write_table}
\alias{db_write_table}
\title{Generic function to post data into Atlas databases}
\usage{
db_write_table(
  table_name,
  data,
  ...,
  schema = "public",
  .host = ATLAS_API_V4_HOST(),
  .token = ATLAS_API_TOKEN(),
  .page_limit = 50000,
  .header = list()
)
}
\arguments{
\item{table_name}{\code{character}. Name of the atlas data object destination.
Corresponds to the name of a within Atlas Postgresql database.}

\item{data}{\code{list} or \code{data.frame}. Either a single record as a \code{list} with
names as attributes or multiple records as a \code{data.frame} with attributes as
columns with attibutes corresponding to the destination column tables.}

\item{...}{\code{character} or \code{numeric} scalar or vector.
Additional parameters to provide to the request. May correspond to postgrest
http request parameters and syntax.}

\item{schema}{\code{character} Schema from the database where is located the data
object is located. Can be either \code{public}, \code{api}, \code{atlas_api} or \code{indicators}.}

\item{.host}{\code{character} Atlas API host url}

\item{.token}{\code{character} Bearer token providing access to the web api}

\item{.page_limit}{\code{integer} Count of objects returned through pagination}

\item{.header}{\code{list} Additional headers to provide to the request.}

\item{.cores}{\code{integer} default \code{4}. Number of cores used to parallelize and
improve rapidity}
}
\description{
Data records attributes (list names or columns) must corresponds to the
table columns it is sent to.
(ie. taxa, observations, datasets, etc).
}
\details{
This function is designed to interface with a web API deployed with PostgREST
}
\keyword{internal}
