% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\docType{data}
\name{dlmalias}
\alias{dlmalias}
\alias{dlm_ll}
\alias{dlm_grades}
\title{DLM Aliases}
\format{
An object of class \code{character} of length 7.

An object of class \code{character} of length 14.
}
\usage{
dlm_ll

dlm_grades
}
\description{
R variable aliases for commonly used vector and values.
}
\details{
\code{dlm_ll} contains all linkage level names for ELA, mathematics, and science.

\code{dlm_grades} contains all DLM grade levels.
}
\keyword{datasets}
