% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-format.R
\name{fmt_table}
\alias{fmt_table}
\title{Title}
\usage{
fmt_table(df, dec_dig = 1, prop_dig = 3, corr_dig = 3, output = NULL)
}
\arguments{
\item{output}{}
}
\value{

}
\description{
Title
}
\examples{

}
\seealso{
Other formatters: 
\code{\link{formatting}},
\code{\link{padding}}
}
\concept{formatters}
