% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-format.R
\name{fmt_table}
\alias{fmt_table}
\title{Center and Decimal Align Tables}
\usage{
fmt_table(df, dec_dig = 1, prop_dig = 3, corr_dig = 3, output = NULL)
}
\arguments{
\item{df}{A data frame or tibble to be formatted for printing in output.}

\item{dec_dig}{The number of decimal places to include for numbers, e.g.,
\code{dec_dig = 1} for 16.5.}

\item{prop_dig}{The number of decimal places to include for numbers bounded
between [0,1], e.g., \code{prop_dig = 2} for .35.}

\item{corr_dig}{The number of decimal places to include for numbers bounded
between [-1,1], e.g., \code{corr_dig = 3} for .205.}

\item{output}{The output format of the table. One of "latex" or "html".
Automatically pulled from document output type if not specified.}
}
\value{
A tibble with the same rows and columns as \code{df}, with numbers
formatted consistently and padded for alignment when printed.
}
\description{
Automatic formatting for tables that should "just work" for most use cases.
For more fine-grained control, see \link[ratlas:formatting]{ratlas::formatting} and
\link[ratlas:padding]{ratlas::padding}.
}
\seealso{
Other formatters: 
\code{\link{formatting}},
\code{\link{padding}}
}
\concept{formatters}
