% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observations.R
\name{get_observations}
\alias{get_observations}
\title{Get observations}
\usage{
get_observations(id = NULL, year = NULL, id_taxa = NULL, ...)
}
\arguments{
\item{id}{Optional. \code{integer} scalar or vector. Returns a dataframe for the
observation with the specified id}

\item{year}{Optional. \code{integer} scalar or vector. Returns a dataframe for
the observations related to the \code{id_taxa}}

\item{id_taxa}{Optional. \code{integer} scalar or vector. Returns a dataframe
for the observations related to the \code{id_taxa}}

\item{...}{Optional. scalar or vector. Returns a dataframe filtered by the
atlas \code{observations} table columns specified as parameter
(ie. \code{id_datasets}, \code{id_variables}, \code{month_obs}, \code{day_obs})}

\item{.cores}{Optional. \code{integer} default \code{4}. Number of cores used to
parallelize and improve rapidity}
}
\value{
\code{tibble} with rows associated with Atlas observations
}
\description{
The function downloads and returns a list or dataframe containing
observations with a single record per row. The function returns a dataframe
containing all records if no parameters are specified (STRONGLY DISCOURAGED
DUE TO LARGE AMOUNT OF OBSERVATIONS CONTAINED WITHIN ATLAS).
The function filters returned records by attributes corresponding to atlas
table columns specified as parameters (ie. \code{id}, \code{year_obs}, \code{id_taxa},
\code{id_datasets}, \code{id_variables}, \code{etc}) with accepted values either being
scalar or vector for single or multiple records.
}
\examples{
# Returns all observations from 2010 to 2015 for taxa
"Cyanocitta cristata" (`id_taxa` = 188) from ebird datasets
(`id_datasets` = 55:102)
obs <- get_observations(
 id_taxa = 188, year = 2010:2015, id_datasets = 55:102)

}
