% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxa.R
\name{get_taxa}
\alias{get_taxa}
\title{Get taxa}
\usage{
get_taxa(id = NULL, scientific_name = NULL, ...)
}
\arguments{
\item{id}{Optional. \code{char} scalar or vector. Returns a dataframe for the
taxon with the specified scientific name}

\item{...}{Optional. scalar or vector. Returns a dataframe filtered for the
parameters specified as columns}
}
\description{
The function downloads and returns a list or dataframe containing a taxonomic
description for each requested species where a row corresponds to an
individual species. The function returns a dataframe for all species if no
options are specified.
}
