% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{get_timeseries}
\alias{get_timeseries}
\title{Get timeseries}
\usage{
get_timeseries(id, id_taxa)
}
\arguments{
\item{id}{Returns a dataframe for the population with the specified id}

\item{id_taxa}{Returns a dataframe containing time series with species
specified by id_taxa.}
}
\value{
A list or dataframe with rows corresponding to individual population.
}
\description{
The function downloads and returns a spatial feature (sf) dataframe
containing time-series, for Quebec, where each row corresponds to an
individual population.
}
