% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/only-if.R
\name{only_if}
\alias{only_if}
\title{Conditionally include steps in a tidyverse pipeline}
\usage{
only_if(condition)
}
\arguments{
\item{condition}{Logical condition to be evaluated}
}
\value{
None. Called for side effects.
}
\description{
Adverb for conditionally skipping steps in a piped workflow.
}
\examples{
tibble::as_tibble(mtcars) |>
  only_if(TRUE)(dplyr::filter)(mpg > 25)

tibble::as_tibble(mtcars) |>
  only_if(FALSE)(dplyr::filter)(mpg > 25)
}
\author{
David Robinson, https://twitter.com/drob/status/785880369073500161
}
