% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rat_cap_words.R
\name{rat_cap_words}
\alias{rat_cap_words}
\title{Capitalization function}
\usage{
rat_cap_words(x)
}
\arguments{
\item{x}{a character string}
}
\description{
This function allows you to capitalize the first letter of every word in a character string.
}
\examples{
rat_cap_words("how are you?")
}
\keyword{capitalization}
