% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratlas.R
\docType{package}
\name{ratlas-package}
\alias{ratlas-package}
\alias{ratlas}
\title{ratlas:}
\description{
The ratlas package serves three main purposes:
}
\details{
\itemize{
\item Project templates for topic guides and technical reports
\item Functions for formatting
\item Consistent ggplot2 themes
}
}
