% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-hrbrthemes.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{scale_x_comma}
\alias{scale_y_comma}
\alias{scale_x_percent}
\alias{scale_y_percent}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{hrbrthemes}{\code{\link[hrbrthemes:scale_x_percent]{scale_x_comma}}, \code{\link[hrbrthemes]{scale_x_percent}}, \code{\link[hrbrthemes:scale_x_percent]{scale_y_comma}}, \code{\link[hrbrthemes:scale_x_percent]{scale_y_percent}}}
}}

