% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okabeito-color-scales.R
\docType{data}
\name{palette_okabeito}
\alias{palette_okabeito}
\alias{scale_colour_okabeito}
\alias{scale_color_okabeito}
\alias{scale_fill_okabeito}
\alias{scale_okabeito}
\title{Okabe-Ito color scale}
\format{
\code{palette_okabeito} is a character vector of 7 hex code colors.
}
\usage{
palette_okabeito

scale_colour_okabeito(aesthetics = "colour", ...)

scale_color_okabeito(aesthetics = "colour", ...)

scale_fill_okabeito(aesthetics = "fill", ...)

scale_okabeito(
  aesthetics,
  use_black = FALSE,
  order = 1:8,
  darken = 0,
  alpha = NA,
  ...
)
}
\arguments{
\item{aesthetics}{The names of the aesthetics that this scale works with.}

\item{...}{Additional arguments passed to \link[ggplot2:discrete_scale]{ggplot2::discrete_scale}.}

\item{use_black}{If \code{TRUE}, scale includes black, otherwise includes gray.}

\item{order}{Numeric vector listing the order in which the colors should be
used. Default is 1:8.}

\item{darken}{Relative amount by which the scale should be darkened (for
positive values) or lightened (for negative values).}

\item{alpha}{Alpha transparency level of the color. Default is no
transparency.}

\item{option}{The color palette to use. One of "okabeito" or "datawrapper".}
}
\value{
A color scale for use in plots created with \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.
}
\description{
The Okabe-Ito color palette is designed to be friendly to colorblind and
colorweak readers. Additionally, this palette should be friendly to
grayscale printing. For details on the palette, see Okabe and Ito (2008).
}
\examples{
scales::show_col(palette_okabeito)

ggplot(penguins, aes(flipper_len, bill_len, color = species)) +
  geom_point(na.rm = TRUE) +
  scale_color_okabeito()

ggplot(penguins, aes(flipper_len, fill = species)) +
  geom_density(na.rm = TRUE, alpha = 0.7) +
  scale_fill_okabeito(order = c(1, 3, 5))
}
\references{
Okabe M., & Ito, K. (2008). \emph{Color universal design (CUD): How to
make figures and presentations that are friendly to Colorblind people}.
Jfly. \url{https://jfly.uni-koeln.de/color/}
}
\keyword{datasets}
