% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{slides_html}
\alias{slides_html}
\title{Create an HTML Slide Deck with R Markdown}
\usage{
slides_html(...)
}
\arguments{
\item{...}{Arguments to be passed to \verb{[xaringan::moon_reader]}}
}
\value{
A modified \code{mood_reader} with ATLAS branding applied.
}
\description{
This is a function called in the output of the YAML of the Rmd file to
specify using the standard DLM tech report pdf document formatting.
}
\examples{
\dontrun{
  output: ratlas::slides_html
}
}
