% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-atlas.R
\name{theme_atlas}
\alias{theme_atlas}
\title{ATLAS ggplot2 theme for consistent graphics}
\usage{
theme_atlas(
  base_size = 11.5,
  base_family = "Archivo Narrow",
  title_size = 18,
  subtitle_size = 13,
  axis_text_size = 10,
  plot_margin = margin(3, 10, 0, 10),
  ...
)
}
\arguments{
\item{base_size}{Base font size, given in pts.}

\item{base_family}{Base font family.}

\item{title_size}{Title font size, given in pts.}

\item{subtitle_size}{Subtitle font size, given in pts.}

\item{axis_text_size}{Axis labels font size, given in pts.}

\item{plot_margin}{Plot margin (specify with \code{\link[ggplot2:element]{ggplot2::margin()}}).}

\item{...}{Additional arguments passed to \code{\link[ggplot2:ggtheme]{ggplot2::theme_minimal()}}.}
}
\value{
A theme for use in plots created with \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.
}
\description{
ATLAS ggplot2 theme for consistent graphics
}
\examples{
ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  labs(x = "Fuel efficiency (mpg)", y = "Weight (tons)",
       title = "Seminal ggplot2 scatterplot example",
       subtitle = "A plot that is only useful for demonstration purposes",
       caption = "Brought to you by the letter 'g'") +
  theme_atlas()

dplyr::count(mtcars, cyl) |>
  dplyr::mutate(cyl = factor(cyl)) |>
  ggplot(aes(cyl, n)) +
  geom_col() +
  geom_text(aes(label = n), nudge_y = 1) +
  labs(x = "Cylinders", y = "Cars",
       title =  "**Seminal *ggplot2* bar chart example**",
       subtitle = "A plot that is only useful for demonstration purposes",
       caption = "Brought to you by the letter 'g'") +
  theme_atlas(base_family = "sans") +
  theme(axis.text.y = element_blank())
}
